+++ src/TypedPersistence/Json/Loading.fs
@@ -1,7 +1,7 @@
 namespace TypedPersistence.Json
 
+open FsToolkit.ErrorHandling
 open System.IO
-open TypedPersistence.Core
 
 open TypedPersistence.Json.Helper
 open TypedPersistence.Json.Types
@@ -9,20 +9,23 @@ open TypedPersistence.Json.Types
 [<AutoOpen>]
 module Loading =
     let private getFileContent filepath =
-        if File.Exists filepath then
-            File.ReadAllText filepath |> Some
-        else
-            None
+        async {
+            if File.Exists filepath then
+                let! content = File.ReadAllTextAsync filepath |> Async.AwaitTask
+                return content |> Some
+            else
+                return None
+        }
 
     let getVersion (filepath: string) =
-        opt {
+        asyncOption {
             let! content = getFileContent filepath
             let! parsed = deserializeJson<OnlyVersion> content
             return parsed.version
         }
 
     let load<'a> (filepath: string) =
-        opt {
+        asyncOption {
             let! content = getFileContent filepath
             let! parsed = deserializeJson<VersionAndData<'a>> content
             return parsed.data