+++ tests/TypedPersistence.UnitTests/Tests.fs
@@ -11,18 +11,25 @@ let testFileDir = "test-files"
 
 type BasicRecord<'a> = { data: 'a }
 
+let testPropertyAsync name func =
+    testProperty name (func >> Async.RunSynchronously)
+
 /// Function to create a basic save and load test for a given type and provider with context
-let genericSaveLoadTest<'a, 'b when 'a: equality> (provider: IPersistenceProvider<'b>)
-                                                  context
-                                                  =
+let genericSaveLoadTest<'a, 'b when 'a: equality>
+    (provider: IPersistenceProvider<'b>)
+    context
+    =
     let typeName = (typeof<'a>).FullName
 
     let context = context typeName
 
-    testProperty ("Save and load: " + typeName) (fun (value: 'a) ->
-        provider.save context value
-        let valueLoaded = (provider.load<'a> context).Value
-        valueLoaded = value)
+    testPropertyAsync ("Save and load: " + typeName) (fun (value: 'a) ->
+        async {
+            do! provider.save context value
+            let! content = provider.load<'a> context
+            let valueLoaded = content.Value
+            return valueLoaded = value
+        })
 
 let basicSaveLoadTests<'a when 'a: equality> () =
     // Test for JsonProvider
@@ -31,8 +38,7 @@ let basicSaveLoadTests<'a when 'a: equality> () =
 
         let context =
             fun typeName ->
-                "SaveLoad"
-                + typeName
+                "SaveLoad" + typeName
                 |> System.Text.Encoding.ASCII.GetBytes
                 |> System.Security.Cryptography.MD5.Create().ComputeHash
                 |> Seq.map (fun c -> c.ToString("X2"))
@@ -48,10 +54,10 @@ let tests =
     testList
         "Test List"
         [ testList
-            "Misc"
-              [ test "Load nonexisting file" {
+              "Misc"
+              [ testAsync "Load nonexisting file" {
                     // TODO: use provider to support new provider
-                    let value = Json.Loading.load<string> "lalelu.txt"
+                    let! value = Json.Loading.load<string> "lalelu.txt"
 
                     Expect.equal value None "Load on non-existing file should return None"
                 } ]
@@ -69,28 +75,31 @@ let tests =
                |> List.concat)
           testList
               "Versioning"
-              [ testProperty "Load version" (fun (version: Version) ->
-                    // TODO: use provider to support new provider
-                    let filepath =
-                        Path.Combine(testFileDir, "load_version.test.json")
+              [ testPropertyAsync "Load version" (fun (version: Version) ->
+                    async {
+                        // TODO: use provider to support new provider
+                        let filepath =
+                            Path.Combine(testFileDir, "load_version.test.json")
 
-                    Json.Saving.saveVersion filepath version "Hello World"
-                    let versionLoaded = (Json.Loading.getVersion filepath).Value
-                    versionLoaded = version)
-                test "Load by version" {
-                    let filepath =
-                        Path.Combine(testFileDir, "load_by_version.test.json")
+                        do! Json.Saving.saveVersion filepath version "Hello World"
+                        let! content = Json.Loading.getVersion filepath
+                        let versionLoaded = content.Value
+                        return versionLoaded = version
+                    })
+                testAsync "Load by version" {
+                    let filepath = Path.Combine(testFileDir, "load_by_version.test.json")
 
                     let data = { data = "Hello World" }
 
-                    Json.Saving.saveVersion filepath 3u data
+                    do! Json.Saving.saveVersion filepath 3u data
 
-                    let value =
-                        match Json.Loading.getVersion filepath with
-                        | Some 3u ->
-                            (Json.Loading.load<BasicRecord<string>> filepath).Value
-                        | Some _
-                        | None -> failwith "Could not load correct version!"
+                    let! version = Json.Loading.getVersion filepath
 
-                    Expect.equal value data "Data should be equal"
+                    match version with
+                    | Some 3u ->
+                        let! content = Json.Loading.load<BasicRecord<string>> filepath
+
+                        Expect.equal content.Value data "Data should be equal"
+                    | Some _
+                    | None -> failwith "Could not load correct version!"
                 } ] ]